The classes "VectorAnopheles" and "HostCategory" have been changed to use now the parameters "proportionMosqLaidEggsSameDay","probabilitySurvivalFeedingCycle", "humanBloodIndex" and "realtiveEntoAvailability".
The Schema Version has been updated to version 17, to comply with the structural
changes.

The SchemaTranslator has been updated to change the parameters for vector using
scenarii. 

The scenarii have been updated, but VecTest and VecFullTest will fail. In Case
of VecFullTest, it maybe so because the paramater EntoAvailability in the old
scenario should be 0.12 instead of 0.012... We have to test it.  



git-svn-id: https://openmalaria.googlecode.com/svn@767 aa1ca794-fd11-11dd-ac65-f74f249cfaac