Moved continuous intervention deployment to before update.
This means humans are seen as being one time-step younger when interventions
are deployed, though were the deployment age reduced accordingly the effect
should still be the same (acting on the next timestep). It seems reasonable to
consider the new behaviour correct: that interventions have their first effect
on humans the timestep *after* they are deployed, so I'm allowing this slight
change in behaviour.

Yes, outputs do change by a small but in some cases significant amount. This is
reduced if age of deployment is reduced by one timestep to counteract change.


git-svn-id: https://openmalaria.googlecode.com/svn@1970 aa1ca794-fd11-11dd-ac65-f74f249cfaac