Changed second "propInfected" to "propInfectious".
Changed the following text:
"The user should be aware that the dynamic transmission modelled is not forced and often fluctuates to a greater or lesser extent (usually below the level input), particularly when close to the limit of sustainable transmission (e.g. with high case-management coverage). Running scenarios without interventions shows the dynamics of EIR and human infectiousness (the continuous output mode is convenient for viewing this)."

git-svn-id: https://openmalaria.googlecode.com/svn@1999 aa1ca794-fd11-11dd-ac65-f74f249cfaac